feat: add JSSG execution service and step executor
- Implemented `JssgExecutionService` for executing JavaScript AST grep tasks.
- Added `StepExecutor` to handle various step actions including script execution, template usage, and codemod execution.
- Enhanced error handling and logging throughout the execution process.
- Integrated capabilities and semantic analysis into the execution flow.
- Updated library structure to include new modules for better organization.
