+++ crates/core/src/lib.rs
@@ -6,10 +6,12 @@ pub mod execution;
 pub(crate) mod execution_stats;
 pub mod file_ops;
 pub mod git_ops;
+pub(crate) mod jssg_execution_service;
 pub(crate) mod periodic;
 pub mod registry;
 pub mod report;
 pub mod shard;
+pub(crate) mod step_executor;
 pub mod structured_log;
 pub mod utils;
 pub mod workflow_runtime;
